Update rag_agent.py
Added a way to load multiple docs when starting the app. the files will be loaded to the vector storage and the vector storage is saved locally between each time you run the file. So if you added a file quit the program and ran it again the file is still in the vector storage.